+++ react-components/src/App.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { BrowserRouter } from 'react-router-dom';
-import Header from './components/Header';
+import { Header } from './components/Header';
 import AppRouter from './router/AppRouter';
 
 export default class App extends Component {
