+++ react-components/src/Icons/EyeIcon.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-export default class EyeIcon extends Component {
+export class EyeIcon extends Component {
   render() {
     return (
       <svg className="svg-icon" viewBox="0 0 1027 1024" xmlns="http://www.w3.org/2000/svg">
