+++ react-components/src/Icons/index.ts
@@ -1,0 +1,4 @@
+export * from './EyeIcon';
+export * from './LikeIcon';
+export * from './SearchIcon';
+export * from './StarIcon';
