+++ react-components/src/components/Card/CardFooter.tsx
@@ -1,22 +1,20 @@
 import React, { Component } from 'react';
-import EyeIcon from '../Icons/EyeIcon';
-import LikeIcon from '../Icons/LikeIcon';
-import StarIcon from '../Icons/StarIcon';
+import * as Icons from '../../Icons';
 
 export default class CardFooter extends Component {
   render() {
     return (
       <div className="card__footer">
         <div className="card__control">
-          <LikeIcon />
+          <Icons.LikeIcon />
           {Math.round(Math.random() * 1000)}
         </div>
         <div className="card__control">
-          <EyeIcon />
+          <Icons.EyeIcon />
           {Math.round(Math.random() * 1000)}
         </div>
         <div className="card__star">
-          <StarIcon />
+          <Icons.StarIcon />
         </div>
       </div>
     );
