+++ react-components/src/components/Card/index.ts
@@ -1,0 +1,1 @@
+export * from './Card';
