+++ react-components/src/components/CardsList/index.ts
@@ -1,0 +1,1 @@
+export * from './CardsList';
