+++ react-components/src/components/Header/Header.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { NavLink } from 'react-router-dom';
 
-export default class Header extends Component {
+export class Header extends Component {
   render() {
     return (
       <header className="header">
