+++ react-components/src/components/Header/index.ts
@@ -1,0 +1,1 @@
+export * from './Header';
