+++ react-components/src/components/SearchBar/SearchBar.test.tsx
@@ -1,5 +1,5 @@
 import { fireEvent, render } from '@testing-library/react';
-import SearchBar from './SearchBar';
+import { SearchBar } from './SearchBar';
 
 test('There is a searchbar on the page', () => {
   const { getByPlaceholderText } = render(<SearchBar />);
