+++ react-components/src/pages/About/About.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-export default class About extends Component {
+export class About extends Component {
   render() {
     return (
       <div className="about">
