+++ react-components/src/pages/About/index.ts
@@ -1,0 +1,1 @@
+export * from './About';
