+++ react-components/src/pages/Home/Home.test.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import Home from './Home';
+import { Home } from './Home';
 
 test('There is a searchbar on the home page', () => {
   const { getByPlaceholderText } = render(<Home />);
