+++ react-components/src/pages/Home/Home.tsx
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
-import CardsList from '../components/CardsList';
-import SearchBar from '../components/SearchBar';
+import { CardsList } from '../../components/CardsList';
+import { SearchBar } from '../../components/SearchBar';
 
-export default class Home extends Component {
+export class Home extends Component {
   render() {
     return (
       <div className="home">
