+++ react-components/src/pages/Page404/Page404.test.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-import Page404 from './Page404';
+import { Page404 } from './Page404';
 
 test('There is a "Not found" text on the 404 page', () => {
   const { getByText } = render(
