+++ react-components/src/pages/Page404/Page404.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 
-export default class Page404 extends Component {
+export class Page404 extends Component {
   render() {
     return (
       <div className="page-404">
