+++ react-components/src/pages/Page404/index.ts
@@ -1,0 +1,1 @@
+export * from './Page404';
